requirements.txt
Django~=3.2
Pillow~=8.4.0
django-admin-sortable2~=1.0
django-tinymce~=3.4.0
environs~=9.5.0
requests~=2.28.1